requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0

# 网页爬取
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
playwright>=1.40.0

# AI集成（DeepSeek API - OpenAI兼容）
openai>=1.10.0

# 视频处理（注意：需要 MoviePy 2.x）
moviepy>=2.0.0
pillow>=10.2.0

# 图片去水印（LaMa模型）
simple-lama-inpainting>=0.1.0
opencv-python-headless>=4.8.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.2
numpy>=1.24.0

# 可选依赖
# edge-tts>=6.1.10      # TTS语音合成（当前未启用）
# redis>=5.0.0           # 缓存去重（可选）
# sqlalchemy>=2.0.0      # 数据库存储（可选）
